requirements.txt
Django=3.1
djangorestframework=3.12.2
psycopg2=2.8.6
sorl-thumbnail=12.7.0
djoser=2.1.0
